+++ src/model-list.js
@@ -16,4 +16,6 @@ export default [
   'track',
   'fragment',
   'playback-rate',
+  'picture-quality',
+  'rotate',
 ];