+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 0.2.3(2018-03-07)
+## 0.2.3(2018-03-27)
 
 #### Bug Fix
 
-无
+* [fix setting of controlbar tooltip position problem](https://github.com/dog-days/html5-player/commit/056394069147cf7d2ecd1cf1efd6a653389b7c4f)
+* [fix firefox setting list fall down problem](https://github.com/dog-days/html5-player/commit/34b4e0bc3a3d7f27e8bb2a6b8dbed6205fccf873)
+* [fix custom subtitle problem](https://github.com/dog-days/html5-player/commit/4ecf37147133fd292ec0265e003a656953f2b653)
 
 #### Update
 
@@ -10,7 +12,11 @@
 
 #### New Function
 
-* 添加 props.stretching 属性，详细请看 API。
+* [添加 props.stretching 属性](https://github.com/dog-days/html5-player/commit/79432a0009af5d25409a57df7d0612066db54cc7)
+* [add hls subtitle feature](https://github.com/dog-days/html5-player/commit/f7e5443e20a4462fb8d0440b9caf18b290481461)
+* [add hls quality swiched feature](https://github.com/dog-days/html5-player/commit/e1554a17d0167550aa4954de1c5b7b0a03f7537b)
+* [add rotate feature](https://github.com/dog-days/html5-player/commit/53b3f98146102329b20c2d127d7c8ef9ea8f0870)
+* [add screen capture feature](https://github.com/dog-days/html5-player/commit/77339118ff6b28b3b22ad8e94ffbebc716463ff7)
 
 ## 0.2.2(2018-03-07)